refactor: Dashboard JS split: extract audit.js (per-domain split, round… [ORB-00153]
Dashboard JS split: extract audit.js (per-domain split, round 2)

Planned-By: claude
